feat: Add SQL script to create users table
The provided code changes add an SQL script that creates a table named "users" with the attributes "id", "email", and "name". The "email" attribute is set to be unique and not null. This change is necessary to support the functionality related to user management in the application.

Please note that the commit message should not include any meta information such as issue references, tags, or author names.